requirements.txt
gymnasium>=0.29.0
minigrid>=2.3.0
stable-baselines3>=2.0.0
torch>=2.0.0
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
opencv-python>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
numpy>=1.24.0
pytest>=7.4.0
